refactor(Package): move bowlinggame to domain
