+++ src/main/java/bowling/BowlingApplication.java
@@ -1,6 +1,6 @@
 package bowling;
 
-import bowling.bowlinggame.BowlingGames;
+import bowling.domain.bowlinggame.BowlingGames;
 import bowling.domain.frame.Frame;
 import bowling.view.InputView;
 import bowling.view.ResultView;
